+++ UWGeodynamics/Underworld_extended/__init__.py
@@ -1,5 +1,0 @@
-from __future__ import print_function,  absolute_import
-from ._mesh import FeMesh_Cartesian
-from ._swarm import Swarm
-from ._swarmvariable import SwarmVariable
-from ._meshvariable import MeshVariable