+++ UWGeodynamics/_boundary_conditions.py
@@ -176,6 +176,7 @@ class BoundaryConditions(object):
 
         # Check that the domain actually contains some boundary nodes
         if isinstance(condition, (list, tuple)) and nodes.data.size > 0:
+
             for dim in range(self.field.data.shape[1]):
 
                 # Scalar condition
@@ -217,10 +218,10 @@ class BoundaryConditions(object):
                 if nodes:
                     self._apply_conditions_nodes(condition, nodes)
 
-        if self.condition_type is "Dirichlet":
+        if self.condition_type == "Dirichlet":
             return uw.conditions.DirichletCondition(
                 variable=self.field, indexSetsPerDof=self._indices)
-        elif self.condition_type is "Neumann":
+        elif self.condition_type == "Neumann":
             _neumann_indices = []
 
             # Remove empty Sets
@@ -387,20 +388,16 @@ class VelocityBCs(BoundaryConditions):
 
             return
 
-        # Expect a list or tuple of dimension mesh.dim.
-        # Check that the domain actually contains some boundary nodes
-        # (nodes is not None)
-        if isinstance(condition, (list, tuple)) and nodes.data.size > 0:
-            for dim in range(self.Model.mesh.dim):
-
-                # Inflow Outflow
-                if isinstance(condition[dim], Balanced_InflowOutflow):
-                    obj = condition[dim]
-                    obj.ynodes = self.Model.mesh.data[nodes.data, 1]
-                    obj._get_side_flow()
-                    self.Model.velocityField.data[nodes.data, dim] = (
-                        obj._get_side_flow())
-                    self._add_to_indices(dim, nodes)
+        if isinstance(condition, Balanced_InflowOutflow):
+            # I don't like that, needs to be redesigned...
+            obj = condition
+            obj.ynodes = self.Model.mesh.data[nodes.data, 1]
+            self.Model.velocityField.data[nodes.data, 0] = (
+                obj._get_side_flow())
+            self.Model.velocityField.data[nodes.data, 1] = 0.
+            self._add_to_indices(0, nodes)
+            self._add_to_indices(1, nodes)
+            return
 
         super(VelocityBCs, self)._apply_conditions_nodes(condition, nodes)
 
@@ -647,7 +644,7 @@ class HeatFlowBCs(BoundaryConditions):
             right = self._get_heat_flux(right[0], right[1])
 
         if top:
-            top = self._get_heat_flux(top[0], right[1])
+            top = self._get_heat_flux(top[0], top[1])
 
         if bottom:
             bottom = self._get_heat_flux(bottom[0], bottom[1])